# book-network-ui/src/openapi/openapi.json
{
     "openapi": "3.0.1",
     "info": {
          "title": "OpenApi specification - Book networking",
          "description": "OpenApi documentation for Spring Security example",
          "termsOfService": "Term of service",
          "contact": {
               "name": "dicka",
               "url": "sidicka.blogspot.co.id",
               "email": "dickanirwansyah@gmail.com"
          },
          "license": {
               "name": "License name",
               "url": "sidicka.blogspot.co.id"
          },
          "version": "1.0"
     },
     "servers": [
          {
               "url": "http://localhost:8081/api/v1",
               "description": "Local ENV"
          },
          {
               "url": "https://dickanirwansyah.aws.co.id/api/v1",
               "description": "Prod ENV"
          }
     ],
     "security": [
          {
               "bearerAuth": []
          }
     ],
     "paths": {
          "/feedbacks": {
               "post": {
                    "tags": [
                         "Feedback"
                    ],
                    "operationId": "saveFeedBack",
                    "requestBody": {
                         "content": {
                              "application/json": {
                                   "schema": {
                                        "$ref": "#/components/schemas/FeedBackRequest"
                                   }
                              }
                         },
                         "required": true
                    },
                    "responses": {
                         "200": {
                              "description": "OK",
                              "content": {
                                   "application/json": {
                                        "schema": {
                                             "type": "integer",
                                             "format": "int64"
                                        }
                                   }
                              }
                         }
                    }
               }
          },
...
          "/feedbacks/book/{book-id}": {
               "get": {
                    "tags": [
                         "Feedback"
                    ],
                    "operationId": "findAllFeedBackByBook",
                    "parameters": [
                         {
                              "name": "book-id",
                              "in": "path",
                              "required": true,
                              "schema": {
                                   "type": "integer",
                                   "format": "int64"
                              }
                         },
                         {
                              "name": "page",
                              "in": "query",
                              "required": false,
                              "schema": {
                                   "type": "integer",
                                   "format": "int32",
                                   "default": 0
                              }
                         },
                         {
                              "name": "size",
                              "in": "query",
                              "required": false,
                              "schema": {
                                   "type": "integer",
                                   "format": "int32",
                                   "default": 10
                              }
                         }
                    ],
                    "responses": {
                         "200": {
                              "description": "OK",
                              "content": {
                                   "application/json": {
                                        "schema": {
                                             "$ref": "#/components/schemas/PageResponseFeedBackResponse"
                                        }
                                   }
                              }
                         }
                    }
               }
          },
...
     },
     "components": {
          "schemas": {
               "FeedBackRequest": {
                    "required": [
                         "bookId",
                         "comment"
                    ],
                    "type": "object",
                    "properties": {
                         "note": {
                              "maximum": 5,
                              "minimum": 0,
                              "type": "number",
                              "format": "double"
                         },
                         "comment": {
                              "type": "string"
                         },
                         "bookId": {
                              "type": "integer",
                              "format": "int64"
                         }
                    }
               },
...
               "FeedBackResponse": {
                    "type": "object",
                    "properties": {
                         "note": {
                              "type": "number",
                              "format": "double"
                         },
                         "comment": {
                              "type": "string"
                         },
                         "ownFeedBack": {
                              "type": "boolean"
                         }
                    }
               },
               "PageResponseFeedBackResponse": {
                    "type": "object",
                    "properties": {
                         "content": {
                              "type": "array",
                              "items": {
                                   "$ref": "#/components/schemas/FeedBackResponse"
                              }
                         },
                         "number": {
                              "type": "integer",
                              "format": "int32"
                         },
                         "size": {
                              "type": "integer",
                              "format": "int32"
                         },
                         "totalElements": {
                              "type": "integer",
                              "format": "int64"
                         },
                         "totalPages": {
                              "type": "integer",
                              "format": "int32"
                         },
                         "first": {
                              "type": "boolean"
                         },
                         "last": {
                              "type": "boolean"
                         }
                    }
               },
...
          },
          "securitySchemes": {
               "bearerAuth": {
                    "type": "http",
                    "description": "JWT auth description",
                    "in": "header",
                    "scheme": "bearer",
                    "bearerFormat": "JWT"
               }
          }
     }
}